Moved function into a tools file
